subscriber/requirements.txt
asyncio-mqtt==0.12.1
asyncpg==0.26.0
autopep8==1.7.0
click==8.1.3
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
paho-mqtt==1.6.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycodestyle==2.9.1
python-dotenv==0.20.0
SQLAlchemy==1.4.40
toml==0.10.2
Werkzeug==2.2.2
zipp==3.8.1
